Split material implementation demo components
