+++ demo-material-impl/src/commonMain/kotlin/com/composeunstyled/demo/materialimpl/ModalNavigationDrawer.kt
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2026 Composable Horizons
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+@file:OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
+@file:Suppress("unused", "UNUSED_PARAMETER")
+
+package com.composeunstyled.demo.material3api
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.material3.DrawerDefaults
+import androidx.compose.material3.DrawerState
+import androidx.compose.material3.DrawerValue
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.rememberDrawerState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import kotlin.math.max
+
+@Composable
+fun ModalNavigationDrawer(
+  drawerContent: @Composable () -> Unit,
+  modifier: Modifier = Modifier,
+  drawerState: DrawerState = rememberDrawerState(DrawerValue.Closed),
+  gesturesEnabled: Boolean = true,
+  scrimColor: Color = DrawerDefaults.scrimColor,
+  content: @Composable () -> Unit,
+) {
+  Box(modifier.fillMaxSize()) {
+    content()
+    if (drawerState.isOpen) {
+      Box(Modifier.fillMaxSize().background(scrimColor))
+      Surface(
+        modifier = Modifier.fillMaxHeight().widthIn(max = 360.dp),
+        color = DrawerDefaults.modalContainerColor,
+        contentColor =
+        androidx.compose.material3.contentColorFor(DrawerDefaults.modalContainerColor),
+      ) {
+        drawerContent()
+      }
+    }
+  }
+}
