+++ demo-material-impl/src/commonMain/kotlin/com/composeunstyled/demo/materialimpl/Scaffold.kt
@@ -1,0 +1,87 @@
+/*
+ * Copyright (c) 2026 Composable Horizons
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+@file:OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
+@file:Suppress("unused", "UNUSED_PARAMETER")
+
+package com.composeunstyled.demo.material3api
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.FabPosition
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ScaffoldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun Scaffold(
+  modifier: Modifier = Modifier,
+  topBar: @Composable () -> Unit = {},
+  bottomBar: @Composable () -> Unit = {},
+  snackbarHost: @Composable () -> Unit = {},
+  floatingActionButton: @Composable () -> Unit = {},
+  floatingActionButtonPosition: FabPosition = FabPosition.End,
+  containerColor: Color = MaterialTheme.colorScheme.background,
+  contentColor: Color = androidx.compose.material3.contentColorFor(containerColor),
+  contentWindowInsets: WindowInsets = ScaffoldDefaults.contentWindowInsets,
+  content: @Composable (PaddingValues) -> Unit,
+) {
+  CompositionLocalProvider(LocalContentColor provides contentColor) {
+    Box(modifier.fillMaxSize().background(containerColor)) {
+      Column(Modifier.fillMaxSize()) {
+        topBar()
+        Box(Modifier.weight(1f)) {
+          content(PaddingValues())
+        }
+        bottomBar()
+      }
+      Box(Modifier.align(Alignment.BottomCenter)) { snackbarHost() }
+      Box(
+        Modifier
+          .align(
+            when (floatingActionButtonPosition) {
+              FabPosition.Start -> Alignment.BottomStart
+              FabPosition.Center -> Alignment.BottomCenter
+              else -> Alignment.BottomEnd
+            },
+          )
+          .padding(16.dp),
+      ) {
+        floatingActionButton()
+      }
+    }
+  }
+}
